+++ MPRController.m
@@ -1966,7 +1966,7 @@ static float deg2rad = M_PI/180.0;
 			[views addObject: mprView2];
 			[views addObject: mprView3];
 			
-			for( int i = views.count-1; i >= 0; i--)
+			for( int i = (long)views.count-1; i >= 0; i--)
 			{
 				if( NSEqualRects( [[views objectAtIndex: i] visibleRect], NSZeroRect))
 					[views removeObjectAtIndex: i];
