+++ browserController.m
@@ -13691,7 +13691,7 @@ static NSArray*	openSubSeriesArray = nil;
             if( f%20 == 0)
             {
                 [NSThread currentThread].progress = (float)f / (float)[copyArray count];
-                [NSThread currentThread].status = N2LocalizedSingularPluralCount( copyArray.count-f, NSLocalizedString(@"file", nil), NSLocalizedString(@"files", nil));
+                [NSThread currentThread].status = N2LocalizedSingularPluralCount( (long)copyArray.count-f, NSLocalizedString(@"file", nil), NSLocalizedString(@"files", nil));
                 
                 if( [NSThread currentThread].isCancelled) //The queue is saved as a plist, we can continue later...
                     break;
